requirements.txt
einops
einops-exts
transformers==4.28.1
torch==2.0.1
torchvision
pillow
more-itertools
datasets
braceexpand
webdataset
wandb
nltk
scipy
inflection
sentencepiece==0.1.98
pycocoevalcap
pycocotools
open_clip_torch>=2.16.0
scikit-learn
sentencepiece

numpy==1.26.2
huggingface_hub==0.19.4
datasets==2.15.0
# others
random-word
sentence-transformers
kmeans-pytorch
hydra-core